+++ fiddles/nativeScript/fiddle-0015-GeoLocation/geolocation/app/app.js
@@ -1,0 +1,5 @@
+
+var application = require("tns-core-modules/application");
+
+application.run({ moduleName: "app-root" });
+